Parse multiple Patterns in a case statement
Second part of JEP 443.
Change the parser so that it can parse a list of Patterns in a case.
Change GuardedPattern AST (internal and and API) so that it has a list
of patterns as its child instead of just one.
Modify some tests that use record patterns with a variable
(eg. `MyPoint(int x, int y) p` vs `MyPoint(int x, int y)`),
since this syntax was only ever used in the preview version
of the feature and is no longer allowed.

Signed-off-by: David Thompson <davthomp@redhat.com>