Fix JavadocConverter.toMethodRefParam errors
Loading bigger projects leads to many:
```
java.lang.IllegalArgumentException: Invalid identifier : >boolean<
	at org.eclipse.jdt.core.dom.SimpleName.setIdentifier(SimpleName.java:242)
	at org.eclipse.jdt.core.dom.AST.newSimpleName(AST.java:2934)
	at org.eclipse.jdt.core.dom.JavadocConverter.toMethodRefParam(JavadocConverter.java:922)
```

Previous javacConverter.convertToType already has the SimpleName proper
when it's really a simple name. In the case of PrimitiveType it fixes
the error.
Note that for some reason the following triggers the problem:
```
@see Object#wait(Object,
int)
```
while the single line doesn't:
```
@see Object#wait(Object, int)
```
But in both cases this seems to be redundant and error prone conversion.